It seems that oo backends can't insert a int into a float struct member: The specific implementation of structures (in C# and Java) must match the STAT_FIELDS description.
(I considered changing rpython.module.r_os_stat.specialize_make_stat_result,
but I could not get it to work)

So:
- st_mtime is a lltype.Float all the time.
- we turn it into a lltype.Signed to build the "struct stat" for the C backend
- structures in C# and Java now hold float times.

Tested with the cli backend, I hope I did not break Java.

+++ pypy/translator/jvm/src/pypy/ll_os.java
@@ -496,7 +496,7 @@ public class ll_os implements Constants {
             else {
                 res.setMode(S_IFREG);
                 res.setSize(f.length());
-                res.setMtime((int)f.lastModified());
+                res.setMtime((double)f.lastModified());
             }
             return res;
         }
